Change to better openldap puppet module.
The new module uses cn=config style configuration, which is much better,
but does not support schema import yet. As a workaround, this is currently
done with executing ldapadd to import LDIF files. I also changed the
authentification mechanism to external in these scripts, since this is now
again possible with the better openldap module.

This solves issue #3.