Change provisioning to take LDAP files from oms-core repository
Composer for php was removed.
The ldif files were removed from this repository and are now in the
 other.
The modules are now in the vagrant folder "ignore/<module-name>", such
 that the user can modify them on the host machine and then
 (upon restarting the service) see the modification in real time
 on the machine.
The class aegee_oms_core is now generalised for all (i.e. it is
 aegee_oms_module)
The LDAP setting is run after cloning the repository

This also closes #32, #28, #26, #22, #18