Small fixes about paths and things to invoke
Rename the main file "server.js" on the profiles module.
Thus, update the init script of the puppet provisioner to call that.
Make sure the version of npm is the same as Travis-ci.
Make sure that upon cloning, the folders inside ignore are present.

For the last point, I had to invent a sort of hack: as git doesn't track
 empty folders I had to come out with something, while at the same time
 making sure that the same folders would be empty at the time of puppet
 provisioning. So I tracked the .delme files and made sure that
 pre-puppet, a shell script was run to delete such files. Then the
 folders would be added to the .gitignore file as not relevant for the
 VM project.